clean up: ignore pycache, venv and db

+++ app/schemas.py
@@ -1,13 +1,15 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field, field_validator
 
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=50)
-    amount: float
+    amount: Decimal
     description: str | None = Field(None, max_length=200)
 
     @field_validator("amount")
-    def validate_amount(cls, v: float) -> float:
+    def validate_amount(cls, v: Decimal) -> Decimal:
         if v <= 0:
             raise ValueError("Amount must be greater than zero")
         return v
@@ -21,7 +23,7 @@ class OperationRequest(BaseModel):
 
 class CreateWalletRequest(BaseModel):
     name: str = Field(..., max_length=50)
-    initial_balance: float = 0
+    initial_balance: Decimal = 0
 
     @field_validator("name")
     def name_not_empty(cls, v: str) -> str:
@@ -31,7 +33,7 @@ class CreateWalletRequest(BaseModel):
         return v
     
     @field_validator("initial_balance")
-    def validate_initial_balance(cls, v: float) -> float:
+    def validate_initial_balance(cls, v: Decimal) -> Decimal:
         if v < 0:
             raise ValueError("Initial balance must be a positive number")
         return v
